colour_identifier_ai/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
boto3==1.38.41
botocore==1.38.41
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==45.0.3
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.8
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
motor==3.7.1
msgpack==1.1.1
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.10.18
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pbs-installer==2025.7.12
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-engineio==4.12.2
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.1
scipy==1.16.0
SecretStorage==3.3.3
sendgrid==6.12.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
webcolors==24.11.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
